feat(form): add important data validation and formatting mask
Add CPF, phone and zip code data formatting mask. In addition to checking if e-mail, CPF/CNPJ,
telephone and zip code are entered correctly.
To help fill in the values to be entered by the user
and prevent incorrect values from being added.
The formatting mask was implemented through the
inputmask library and the useRef hook. Validation was implemented with the yup library integrated
with formik.

fixes #4