feat(form): add automatic address filling by zip code
Add functionality to populate the address fields, entering the zip code and verifying that it is
correct.
To facilitate user data entry and provide greater assurance that the data entered is
correct.
The functionality will be implemented using the cep-promise library and the state control
from the formik library.

fixes #4