feat(firebase): add connection with firebase and persistence of registered data
Adds firebase connection and creates functionality to log form data into the database. In addition
to creating real-time data in the table with the possibility of deleting it individually.
So that
you have a persistence of registered data even when the user leaves the page.
This functionality
was implemented through firebase, useState and useEffect.

fixes #6