refactor: refactor the database used and create context to handle processes
refactor the firebase database, replacing realtime with firestone. What is better with offline
operation (add, search and remove). In addition to creating a context to deal with the status of
processes performed in the database.
Allow the app to be used even offline if there is a period of
instability in the connection. In addition to providing feedback to the user about all processes
that occurred related to database operations.
Implementations were performed using firebase
firestone, useState, useEffect and React contextAPI.

fixes #17 and #16